+++ R/cam02ucs.R
@@ -56,16 +56,30 @@ pth_new_cam02ucs <- function(mat, c = 0.69, Y_b = 20, L_A = 64 / pi / 5,
       c = c,
       Y_b = Y_b,
       L_A = L_A,
-      whitepoint = whitepoint,
-      transformer = function(color) {
-        pth_to_cam02ucs(color, c = c, Y_b = Y_b, L_A = L_A, whitepoint = whitepoint)
-      }
+      whitepoint = whitepoint
     )
 
   # attach labels
   result <- label_cols(result, cam02ucs$labels)
 
   result
+}
+
+#' @rdname pth_transformer
+#' @export
+#'
+pth_transformer.pth_cam02ucs <- function(mat, ...) {
+
+  function(color) {
+    pth_to_cam02ucs(
+      color,
+      c = attr(mat, "c"),
+      Y_b = attr(mat, "Y_b"),
+      L_A = attr(mat, "L_A"),
+      whitepoint = attr(mat, "whitepoint")
+    )
+  }
+
 }
 
 #' @export