+++ R/cam16ucs.R
@@ -67,23 +67,31 @@ pth_new_cam16ucs <- function(mat, c = 0.69, Y_b = 20, L_A = 64 / pi / 5,
       Y_b = Y_b,
       L_A = L_A,
       exact_inversion = exact_inversion,
-      whitepoint = whitepoint,
-      transformer = function(color) {
-        pth_to_cam16ucs(
-          color,
-          c = c,
-          Y_b = Y_b,
-          L_A = L_A,
-          exact_inversion = exact_inversion,
-          whitepoint = whitepoint
-        )
-      }
+      whitepoint = whitepoint
     )
 
   # attach labels
   result <- label_cols(result, cam16ucs$labels)
 
   result
+}
+
+#' @rdname pth_transformer
+#' @export
+#'
+pth_transformer.pth_cam16ucs <- function(mat, ...) {
+
+  function(color) {
+    pth_to_cam16ucs(
+      color,
+      c = attr(mat, "c"),
+      Y_b = attr(mat, "Y_b"),
+      L_A = attr(mat, "L_A"),
+      exact_inversion = attr(mat, "exact_inversion"),
+      whitepoint = attr(mat, "whitepoint")
+    )
+  }
+
 }
 
 #' @export