+++ NEWS.md
@@ -88,6 +88,7 @@ New functions to support visualizing the sRGB gamut (also other colorsets) in po
 
   - `pth_mat_gamut()`: Returns representation of sRGB gamut in a given color space.
   - `pth_cvd_grid()`, `pth_cvd_grid_full()`: Helpers to return tibble for possibities for color-vision deficiency and severity.
+  - `pth_transformer()`: Returns a function to transform to the colorpace of the object that generated it.
   
 # colorpath 0.0.0 (development version)